requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.21.0
Django==5.2
django-admin-interface==0.30.0
django-apscheduler==0.7.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-ninja==1.4.1
django-picklefield==3.3
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
jinxed==1.3.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
redis==3.5.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
wcwidth==0.2.13
